requirements.txt
aistudio-sdk==0.3.6
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bce-python-sdk==0.9.45
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.120.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.11
ImageHash==4.3.2
imagesize==1.4.1
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
# keyboard requires root on Linux, so it's Windows-only. pynput is used on Linux instead.
keyboard==0.13.5 ; sys_platform == "win32"
mss==10.1.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
modelscope==1.29.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.0
paddlepaddle==3.2.0
paddlex==3.3.6
pandas==2.3.2
pillow==12.0.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
prettytable==3.16.0
protobuf==6.33.0
psutil==7.1.2
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
# PyGetWindow is Windows/macOS only - not needed on Linux when using pywinctl
PyGetWindow==0.0.9 ; sys_platform == "win32" or sys_platform == "darwin"
pynput==1.8.1
# pywinctl works cross-platform and provides PID-based window detection
pywinctl==0.4.1
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
pypdfium2==5.0.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
PyWavelets==1.8.0
pywin32-ctypes==0.2.3
# pywin32 is Windows-only; use environment marker so Linux installations will ignore it
pywin32==311 ; sys_platform == "win32"
# python-xlib is Linux-only, for X11 window management fallback
python-xlib==0.33 ; sys_platform == "linux"
PyYAML==6.0.2
RapidFuzz==3.14.1
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.14.2
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14